requirements.txt
Office365-REST-Python-Client
python-dotenv
oracledb
schedule
suds
paramiko